fix: Info bug fixes and improvements
